Refactor manager components
