+++ packages/manager-components/src/components/Trade/container.tsx
@@ -1,6 +1,7 @@
 import { withState } from 'recompose';
 import compose from 'recompose/compose';
+import Trade from './index';
 
 const withActiveTabState = withState('activeTabIndex', 'setTabIndex', 0);
 
-export default compose(withActiveTabState);
+export default compose(withActiveTabState)(Trade);
